Crate kalavar_core

Modules

core
A selection of utilities and helper methods to ensure that the program runs smoothyl

core_structures
Structure definitions for things like databases and tables

errors
Custom defined error types used by the program to report clear, concise error messages back to the user so they can get their database up and running ASAP

managers
Structure definitions for all of the management structures used within the program

Statics

ALLOC

Functions

main