Module kalavar_core::compilers::KQL::utilities

The structures required for Queries

Structs

Filter
The structure used for filtering data from a query where possible

Query
A general structure matching a Query after it has been compiled

SubFilter
A sub-structure used for defining filters on content

Variable
A structure defining a variable

Enums

Operation
An enumerator defining the different operations allowed to queries

VariablePrivacy
An enumerator defining the privacy types of variables

VariableType
An enumerator defining the types of variables possible