Function kalavar_core::core::utils::connection_handling::handle

pub async fn handle(_con: Connection)

The following content is undocumented due to not being ready for documentation at this time.

You are welcome to attempt to make sense of it though.