Enum kalavar_core::core_structures::column::ColumnTypeEnum

pub enum ColumnTypeEnum {
    String,
    JSON,
    Bool,
    Byte,
    BLOB,
    Integer8,
    Integer16,
    Integer32,
    Integer64,
    SignedInteger8,
    SignedInteger16,
    SignedInteger32,
    SignedInteger64,
    BigInteger,
    SignedBigInteger,
    Snowflake,
    UUID,
    SonyFlake,
    RGB,
    RGBA,
    CMYK,
    Hex,
    Array,
    Enum,
    IPv4,
    IPv6,
    Mac,
    Mac8,
    Timestamp,
    NaiveTimestamp,
}

An enumerator which defines the available column types I recommend reading the documentation at the link below for details on them all: https://kalavar.cf/documentation/data-types/

Variants

String

A generic String type

JSON

JSON flavoured string

Bool

Boolean data type for columns

Byte

8 Bit storage type

BLOB

Binary Large Object type

Integer8

Unsigned 8 bit integer

Integer16

Unsigned 16 bit integer

Integer32

Unsigned 32 bit integee

Integer64

Unsigned 64 bit integer

SignedInteger8

Signed 8 bit integer

SignedInteger16

Signed 16 bit integer

SignedInteger32

Signed 32 bit integer

SignedInteger64

Signed 64 bit integer

BigInteger

An unsigned integer, comprised of other integers to provide maximum storage capacity

SignedBigInteger

A signed integer, comprised of other integers to provide maximum storage capacity

Snowflake

Snowflake unique identifier

UUID

Universally Unique Identifier

SonyFlake

Sony's re-thought version of Snowflake

RGB

Red, Green, Blue color type

RGBA

Red, Green, Blue, Alpha color type

CMYK

Cyan, Magenta, Yellow, Black color type

Hex

Hexadecimal color type

Array

An array type for the column

Enum

An enumerator type for the column, see the docs for more info

IPv4

Simple validated IPv4 address type

IPv6

Simple validated IPv6 address type

Mac

Simple MAC address type

Mac8

Mac8, currently unknown but PostgreSQL has it so it must be needed by someone

Timestamp

A timestamp, prefixed with information about the timezone it is for to allow for conversion

NaiveTimestamp

A timestamp without the prefixed timezone data, assumes it is UTC and converts to Timetstamp as though it was UTC based, even if it may not be