Struct kalavar_core::core_structures::connection_protocol::ConnectionProtocolMessage

pub struct ConnectionProtocolMessage {
    pub recipient: usize,
    pub sender: usize,
    pub inner_type: DataType,
    pub inner: (Option<String>, Option<Vec<u8>>, Option<(f64, f64, f64)>),
    pub opcode: OpCodes,
}

A structure defined to help move information between threads via broadcast channels

Fields
recipient: usize

The intended recipient of the payload (0 for the protocol thread, or anything else if it is a client connection)

sender: usize

The ID of the thread which sent the payload

inner_type: DataType

The type that should be expected as present

inner: (Option<String>, Option<Vec<u8>>, Option<(f64, f64, f64)>)

The inner data A load of Option types with varying containers for different data structures as needed

opcode: OpCodes

The OP code of the payload, see the docs at the link below for a breakdown: https://kalavar.cf/documentation/general/opcodes

Implementations

impl ConnectionProtocolMessage

The following content is undocumented due to not being ready for documentation at this time.

You are welcome to attempt to make sense of it though.

pub fn new_mem(used: f64, total: f64) -> ConnectionProtocolMessage

pub fn new_con(id: &usize) -> ConnectionProtocolMessage