Module kalavar_core::core_structures::row

Definition information for the row and cell structure

Structs

Cell
A structure defining the type of column that the data is from, as well as the raw bytes of the given data

Row
A structure wrapping the cells of each row in a neat little basket