Module kalavar_core::managers

Structure definitions for all of the management structures used within the program

Modules

analytics
Documentation for the analytics manager

authentication
Documentation for the authentication manager

config
Documentation for the configuration manager

connections
Documentation for connections

encryption
Documentation for the encryption manager

implementors
Documentation for the web UI Implentation definitions for all of the management structures used within the program

logging
Documentation for the logging manager

permission
Documentation for the permissions manager

storage
Documentation for the storage manager