Module kalavar_core::managers::analytics

Documentation for the analytics manager

Structs

Analytics
A sub structure which contains all of the possible information which will be displayed to the user via the statistics viewer utility as well as any information which MIGHT be shared with the developers based on privacy config

AnalyticsManager
The manager which reports analytics information to the user (and back to the core developers when allowed to)

CPUInfo
An information storage structure designed to store information about the CPU

DiskInfo
An information storage structure designed to store information about the disk

InnerMemoryInfo
An information storage structure designed to store information about the memory

MemoryInfo
An information storage structure designed to store information about the memory

OSInfo
An information storage structure designed to store information about the OS