Struct kalavar_core::managers::authentication::AuthManager

pub struct AuthManager {
    pub permissions: PermissionManager,
    pub remote: String,
    pub is_admin: bool,
    pub is_root: bool,
    pub username: String,
    pub certificate: String,
    pub validated: bool,
}

A utility structure for managing authentication with the client

Fields
permissions: PermissionManager

A structure representing the permissions that the user has across all of the system

remote: String

The remote address of the client

is_admin: bool

If the user is considered an administrator of the system

is_root: bool

If the user is considered to be the "root" user

username: String

The user's name

certificate: String

The certificate which the user used during authentication

validated: bool

If the user has been successfully logged in, or has been instantiated outside of a log-on environment