Module kalavar_core::managers::config::post

Structs

Config
A structure containing the root keys of the config.toml file

ConfigManager
A utility structure designed to parse the content of the configuration files used to make Kalavar work.

LanguageConfig
A utility structure to define things relating to language used within the system

LogConfig
Utility structure to define things relating to output logging used within the system

NetConfig
Network configuration option utility structure

PrivacyConfig
A utility structure for defining things related to the privacy of the user

Enums

NamingConvention
An enumerator for making naming convention simpler

PrivacyMode
An enumerator for making privacy checking simpler